+++ R/context.R
@@ -109,11 +109,12 @@ close.nanoContext <- function(con, ...) invisible(.Call(rnng_ctx_close, con))
 #'     in the vector of choices i.e. 1L for 'serial' or 2L for 'raw'.
 #' @param recv_mode [default 'serial'] mode of vector to be received - one of
 #'     'serial', 'character', 'complex', 'double', 'integer', 'logical',
-#'     'numeric', or 'raw'. The default 'serial' means a serialised R object,
-#'     for the other modes, the raw vector received will be converted into the
-#'     respective mode. Alternatively, for performance, specify an integer
-#'     position in the vector of choices e.g. 1L for 'serial', 2L for 'character'
-#'     etc.
+#'     'numeric', 'raw', or 'string'. The default 'serial' means a serialised R
+#'     object, for the other modes, the raw vector received will be converted
+#'     into the respective mode. Note that 'string' is defined here as a character
+#'     scalar and is a faster alternative to 'character' for receiving a single
+#'     string. Alternatively, for performance, specify an integer position in
+#'     the vector of choices e.g. 1L for 'serial', 2L for 'character' etc.
 #' @param timeout [default NULL] integer value in milliseconds or NULL, which
 #'     applies a socket-specific default, usually the same as no timeout. Note
 #'     that this applies to receiving the request. The total elapsed time would