+++ R/sendrecv.R
@@ -83,10 +83,12 @@ send <- function(con, data, mode = c("serial", "raw"), block = NULL)
 #'
 #' @param con a Socket, Context or Stream.
 #' @param mode [default 'serial'] mode of vector to be received - one of 'serial',
-#'     'character', 'complex', 'double', 'integer', 'logical', 'numeric', or 'raw'.
-#'     The default 'serial' means a serialised R object, for the other modes,
-#'     the raw vector received will be converted into the respective mode.
-#'     For Streams, 'serial' is not an option and the default is 'character'.
+#'     'character', 'complex', 'double', 'integer', 'logical', 'numeric', 'raw',
+#'     or 'string'. The default 'serial' means a serialised R object, for the
+#'     other modes, the raw vector received will be converted into the respective
+#'     mode. Note that 'string' is defined here as a character scalar and is a
+#'     faster alternative to 'character' for receiving a single string. For
+#'     Streams, 'serial' is not an option and the default is 'character'.
 #'     Alternatively, for performance, specify an integer position in the vector
 #'     of choices e.g. 1L for 'serial', 2L for 'character' etc.
 #' @param n [default 65536L] applicable to Streams only, the maximum number of
@@ -156,7 +158,7 @@ send <- function(con, data, mode = c("serial", "raw"), block = NULL)
 #'
 recv <- function(con,
                  mode = c("serial", "character", "complex", "double",
-                          "integer", "logical", "numeric", "raw"),
+                          "integer", "logical", "numeric", "raw", "string"),
                  block = NULL,
                  n = 65536L)
   .Call(rnng_recv, con, mode, block, n)